+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 0.9.3
 - `@use-gpu/glyph`: Support loading emoji PNGs on-demand
 - `@use-gpu/layout`: Add `aspect` to elements for sizing w/h relative to each other
-- `@use-gpu/plot`: Add UVs and STs to `<Surface>`
+- `@use-gpu/workbench`: Add UVs and STs to `<LineLayer>`, `<PointLayer>`, `<SurfaceLayer>`
+- `@use-gpu/workbench`: Use input position as default STs
 
 0.9.2
 - `@use-gpu/inspect`: Fix inspecting of large typed arrays